# backend/docs/openapi/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Donguri API
  version: 1.0.0
paths:
  /api/v1/ping:
    get:
      summary: Ping
      operationId: ping
      responses:
        "200":
          description: pong
          content:
            text/plain:
              schema:
                type: string
                example: pong
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/rooms/join:
    post:
      summary: Join matching queue
      operationId: joinRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomRequest"
      responses:
        "202":
          description: Joined or assigned to a room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRoomResponse"
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/rooms/{room_id}/stream:
    get:
      summary: Subscribe room events (SSE)
      operationId: subscribeRoomStream
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: player_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: room_started
                  data: {"room_id":"00000000-0000-0000-0000-000000000001"}

                  event: room_updated
                  data: {"intersection_count":3}
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/rooms/{room_id}:
    get:
      summary: Get room state
      operationId: getRoom
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Player-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Current room and player-related state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomStateResponse"
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/rooms/{room_id}/connections/{target_id}:
    put:
      summary: Upsert connection intent by target player ID
      operationId: putConnection
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: target_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Player-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionRequest"
      responses:
        "200":
          description: Intent accepted or swap resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionResponse"
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/rooms/{room_id}/swaps/intents/{target_player_id}:
    delete:
      summary: Cancel swap intent by target player ID
      operationId: deleteSwapIntent
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: target_player_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Player-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Intent removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSwapIntentResponse"
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/rooms/{room_id}/players/me/intent:
    patch:
      summary: Set directional swap intent ON (prev or next)
      operationId: patchMyDirectionalIntent
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Player-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DirectionalIntentRequest"
      responses:
        "200":
          description: Directional intent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectionalIntentResponse"
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Set directional swap intent OFF (prev or next)
      operationId: deleteMyDirectionalIntent
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Player-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DirectionalIntentRequest"
      responses:
        "200":
          description: Directional intent removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectionalIntentResponse"
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/rooms/{room_id}/messages:
    get:
      summary: Get message history related to myself
      operationId: getMessages
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Player-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Message history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Send message to adjacent player
      operationId: createMessage
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Player-ID
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "default":
          description: Error response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    JoinRoomRequest:
      type: object
      properties:
        name:
          type: string
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
      required:
        - name
        - lat
        - lng

    JoinRoomResponse:
      type: object
      properties:
        room_id:
          type: string
          format: uuid
        player_id:
          type: string
          format: uuid
      required:
        - room_id
        - player_id

    RoomStateResponse:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/RoomStatus"
        time_left:
          type: integer
          minimum: 0
          description: Remaining seconds. null when not started.
          nullable: true
        intersection_count:
          type: integer
          minimum: 0
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        my_intents:
          type: array
          items:
            type: string
            format: uuid
        received_intents:
          type: array
          items:
            type: string
            format: uuid
      required:
        - status
        - intersection_count
        - players
        - my_intents
        - received_intents

    RoomStatus:
      type: string
      enum:
        - matching
        - playing
        - finished

    Player:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        order_index:
          type: integer
          minimum: 0
        location:
          $ref: "#/components/schemas/Location"
      required:
        - id
        - name
        - order_index
        - location

    Location:
      type: object
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
      required:
        - lat
        - lng

    ConnectionRequest:
      type: object
      properties:
        needs_swap:
          type: boolean
      required:
        - needs_swap

    ConnectionResponse:
      type: object
      properties:
        matched:
          type: boolean
          description: true when both sides were ON and swap executed
        room_status:
          $ref: "#/components/schemas/RoomStatus"
      required:
        - matched
        - room_status

    DeleteSwapIntentResponse:
      type: object
      properties:
        removed:
          type: boolean
      required:
        - removed

    DirectionalIntentRequest:
      type: object
      properties:
        direction:
          type: string
          enum:
            - prev
            - next
      required:
        - direction

    DirectionalIntentResponse:
      type: object
      properties:
        direction:
          type: string
          enum:
            - prev
            - next
        enabled:
          type: boolean
      required:
        - direction
        - enabled

    CreateMessageRequest:
      type: object
      properties:
        receiver_id:
          type: string
          format: uuid
        content:
          type: string
          minLength: 1
          maxLength: 2000
      required:
        - receiver_id
        - content

    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        room_id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        receiver_id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - room_id
        - sender_id
        - receiver_id
        - content
        - created_at

    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message