# backend/internal/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn8AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn2AllowedHeaders = map[string]string{
		"GET": "X-Player-Id",
	}
	rn12AllowedHeaders = map[string]string{
		"PUT": "Content-Type,X-Player-Id",
	}
	rn3AllowedHeaders = map[string]string{
		"GET":  "X-Player-Id",
		"POST": "Content-Type,X-Player-Id",
	}
	rn5AllowedHeaders = map[string]string{
		"DELETE": "Content-Type,X-Player-Id",
		"PATCH":  "Content-Type,X-Player-Id",
	}
	rn7AllowedHeaders = map[string]string{
		"DELETE": "X-Player-Id",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [2]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/"

			if l := len("/api/"); len(elem) >= l && elem[0:l] == "/api/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'r': // Prefix: "rooms/"

				if l := len("rooms/"); len(elem) >= l && elem[0:l] == "rooms/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'j': // Prefix: "join"
					origElem := elem
					if l := len("join"); len(elem) >= l && elem[0:l] == "join" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleJoinRoomRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "POST",
								allowedHeaders: rn8AllowedHeaders,
								acceptPost:     "application/json",
								acceptPatch:    "",
							})
						}

						return
					}

					elem = origElem
				}
				// Param: "room_id"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleGetRoomRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: rn2AllowedHeaders,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "connections/"

						if l := len("connections/"); len(elem) >= l && elem[0:l] == "connections/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "target_id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[1] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "PUT":
								s.handlePutConnectionRequest([2]string{
									args[0],
									args[1],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "PUT",
									allowedHeaders: rn12AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'm': // Prefix: "messages"

						if l := len("messages"); len(elem) >= l && elem[0:l] == "messages" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGetMessagesRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "POST":
								s.handleCreateMessageRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "GET,POST",
									allowedHeaders: rn3AllowedHeaders,
									acceptPost:     "application/json",
									acceptPatch:    "",
								})
							}

							return
						}

					case 'p': // Prefix: "players/me/intent"

						if l := len("players/me/intent"); len(elem) >= l && elem[0:l] == "players/me/intent" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleDeleteMyDirectionalIntentRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "PATCH":
								s.handlePatchMyDirectionalIntentRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, notAllowedParams{
									allowedMethods: "DELETE,PATCH",
									allowedHeaders: rn5AllowedHeaders,
									acceptPost:     "",
									acceptPatch:    "application/json",
								})
							}

							return
						}

					case 's': // Prefix: "s"

						if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 't': // Prefix: "tream"

							if l := len("tream"); len(elem) >= l && elem[0:l] == "tream" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "GET":
									s.handleSubscribeRoomStreamRequest([1]string{
										args[0],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "GET",
										allowedHeaders: nil,
										acceptPost:     "",
										acceptPatch:    "",
									})
								}

								return
							}

						case 'w': // Prefix: "waps/intents/"

							if l := len("waps/intents/"); len(elem) >= l && elem[0:l] == "waps/intents/" {
								elem = elem[l:]
							} else {
								break
							}

							// Param: "target_player_id"
							// Leaf parameter, slashes are prohibited
							idx := strings.IndexByte(elem, '/')
							if idx >= 0 {
								break
							}
							args[1] = elem
							elem = ""

							if len(elem) == 0 {
								// Leaf node.
								switch r.Method {
								case "DELETE":
									s.handleDeleteSwapIntentRequest([2]string{
										args[0],
										args[1],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "DELETE",
										allowedHeaders: rn7AllowedHeaders,
										acceptPost:     "",
										acceptPatch:    "",
									})
								}

								return
							}

						}

					}

				}

			case 'v': // Prefix: "v1/ping"

				if l := len("v1/ping"); len(elem) >= l && elem[0:l] == "v1/ping" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handlePingRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, notAllowedParams{
							allowedMethods: "GET",
							allowedHeaders: nil,
							acceptPost:     "",
							acceptPatch:    "",
						})
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [2]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/"

			if l := len("/api/"); len(elem) >= l && elem[0:l] == "/api/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'r': // Prefix: "rooms/"

				if l := len("rooms/"); len(elem) >= l && elem[0:l] == "rooms/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'j': // Prefix: "join"
					origElem := elem
					if l := len("join"); len(elem) >= l && elem[0:l] == "join" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = JoinRoomOperation
							r.summary = "Join matching queue"
							r.operationID = "joinRoom"
							r.operationGroup = ""
							r.pathPattern = "/api/rooms/join"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

					elem = origElem
				}
				// Param: "room_id"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = GetRoomOperation
						r.summary = "Get room state"
						r.operationID = "getRoom"
						r.operationGroup = ""
						r.pathPattern = "/api/rooms/{room_id}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "connections/"

						if l := len("connections/"); len(elem) >= l && elem[0:l] == "connections/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "target_id"
						// Leaf parameter, slashes are prohibited
						idx := strings.IndexByte(elem, '/')
						if idx >= 0 {
							break
						}
						args[1] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "PUT":
								r.name = PutConnectionOperation
								r.summary = "Upsert connection intent by target player ID"
								r.operationID = "putConnection"
								r.operationGroup = ""
								r.pathPattern = "/api/rooms/{room_id}/connections/{target_id}"
								r.args = args
								r.count = 2
								return r, true
							default:
								return
							}
						}

					case 'm': // Prefix: "messages"

						if l := len("messages"); len(elem) >= l && elem[0:l] == "messages" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GetMessagesOperation
								r.summary = "Get message history related to myself"
								r.operationID = "getMessages"
								r.operationGroup = ""
								r.pathPattern = "/api/rooms/{room_id}/messages"
								r.args = args
								r.count = 1
								return r, true
							case "POST":
								r.name = CreateMessageOperation
								r.summary = "Send message to adjacent player"
								r.operationID = "createMessage"
								r.operationGroup = ""
								r.pathPattern = "/api/rooms/{room_id}/messages"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					case 'p': // Prefix: "players/me/intent"

						if l := len("players/me/intent"); len(elem) >= l && elem[0:l] == "players/me/intent" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "DELETE":
								r.name = DeleteMyDirectionalIntentOperation
								r.summary = "Set directional swap intent OFF (prev or next)"
								r.operationID = "deleteMyDirectionalIntent"
								r.operationGroup = ""
								r.pathPattern = "/api/rooms/{room_id}/players/me/intent"
								r.args = args
								r.count = 1
								return r, true
							case "PATCH":
								r.name = PatchMyDirectionalIntentOperation
								r.summary = "Set directional swap intent ON (prev or next)"
								r.operationID = "patchMyDirectionalIntent"
								r.operationGroup = ""
								r.pathPattern = "/api/rooms/{room_id}/players/me/intent"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					case 's': // Prefix: "s"

						if l := len("s"); len(elem) >= l && elem[0:l] == "s" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case 't': // Prefix: "tream"

							if l := len("tream"); len(elem) >= l && elem[0:l] == "tream" {
								elem = elem[l:]
							} else {
								break
							}

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "GET":
									r.name = SubscribeRoomStreamOperation
									r.summary = "Subscribe room events (SSE)"
									r.operationID = "subscribeRoomStream"
									r.operationGroup = ""
									r.pathPattern = "/api/rooms/{room_id}/stream"
									r.args = args
									r.count = 1
									return r, true
								default:
									return
								}
							}

						case 'w': // Prefix: "waps/intents/"

							if l := len("waps/intents/"); len(elem) >= l && elem[0:l] == "waps/intents/" {
								elem = elem[l:]
							} else {
								break
							}

							// Param: "target_player_id"
							// Leaf parameter, slashes are prohibited
							idx := strings.IndexByte(elem, '/')
							if idx >= 0 {
								break
							}
							args[1] = elem
							elem = ""

							if len(elem) == 0 {
								// Leaf node.
								switch method {
								case "DELETE":
									r.name = DeleteSwapIntentOperation
									r.summary = "Cancel swap intent by target player ID"
									r.operationID = "deleteSwapIntent"
									r.operationGroup = ""
									r.pathPattern = "/api/rooms/{room_id}/swaps/intents/{target_player_id}"
									r.args = args
									r.count = 2
									return r, true
								default:
									return
								}
							}

						}

					}

				}

			case 'v': // Prefix: "v1/ping"

				if l := len("v1/ping"); len(elem) >= l && elem[0:l] == "v1/ping" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = PingOperation
						r.summary = "Ping"
						r.operationID = "ping"
						r.operationGroup = ""
						r.pathPattern = "/api/v1/ping"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}